Debugging performance

+++ official/resnet/imagenet_main.py
@@ -137,7 +137,7 @@ def synthetic_input_fn(is_training, data_dir, batch_size, num_epochs=1):
   input_shape = [_DEFAULT_IMAGE_SIZE, _DEFAULT_IMAGE_SIZE, _NUM_CHANNELS]
   feature_vals = tf.random_normal(num_examples + input_shape)
   label_vals = tf.random_uniform(
-      num_examples, minval=0, maxval=_NUM_CLASSES - 1)
+      num_examples, minval=0, maxval=_NUM_CLASSES - 1, dtype=tf.int32)
   label_vals = tf.one_hot(label_vals, _NUM_CLASSES)
 
   dataset = tf.data.Dataset.from_tensor_slices((feature_vals, label_vals))
